Add react-router-doom, basic routing home
