+++ src/index.js
@@ -1,14 +1,24 @@
+//** Libaries */
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import './index.scss';
+import { BrowserRouter } from 'react-router-dom';
+
+//** Components*/
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+
+//** Css*/
+import './index.scss';
+
+
 
 const container = document.getElementById('root');
 const root = createRoot(container);
 root.render(
   <React.StrictMode>
-    <App />
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </React.StrictMode>
 );
